+++ .env.production
@@ -1,0 +1,2 @@
+REACT_APP_API='https://api.icetea.io'
+REACT_APP_INIT_BLOCK=1
