+++ src/service/api/base-api.js
@@ -1,8 +1,16 @@
 import axios from 'axios';
-import {host} from '../../evironment/env';
+
+export const listBlocks = '/block/list';
+export const countBlock = '/block/count';
+export const singleBlock = '/block'; //:height
+export const latestBlock = '/block/latest';
+
+export const listTxs = '/tx/list';
+export const singleTx = '/tx'; //:hash
+export const countTxs = '/tx/count';
 
 export const _get = async (params, api) => {
-    let url = host + api;
-    let response = await axios.get(url, {params , headers: null});
-    return response;
-}
+  const url = process.env.REACT_APP_API + api;
+  const response = await axios.get(url, { params, headers: null });
+  return response;
+};
