--- src/service/api/list-api.js
@@ -1,8 +1,0 @@
-export const listBlocks = "/block/list";
-export const countBlock = "/block/count";
-export const singleBlock = "/block"; //:height
-export const latestBlock = "/block/latest";
-
-export const listTxs = "/tx/list";
-export const singleTx = "/tx"; //:hash
-export const countTxs = "/tx/count";
