+++ src/service/api/transactions.js
@@ -1,35 +1,6 @@
 import store from '../../store';
 import * as actions from '../../store/actions';
-import { _get } from './base-api';
-import { listBlocks, listTxs, countBlock, countTxs, singleBlock } from './list-api';
-
-export const getListBlockApi = async params => {
-  const response = await _get(params, listBlocks);
-
-  try {
-    if (response.status === 200) {
-      const { data } = response;
-      //   console.log("getListBlockApi", response);
-      store.dispatch(actions.setBlocks(data));
-    }
-  } catch (err) {
-    throw err;
-  }
-};
-
-export const getTotalBlockApi = async params => {
-  const response = await _get(params, countBlock);
-
-  try {
-    if (response.status === 200) {
-      const { data } = response;
-      // console.log("getListBlockApi", response);
-      store.dispatch(actions.setTotalBlock(data[0] ? data[0].count : 0));
-    }
-  } catch (err) {
-    throw err;
-  }
-};
+import { _get, listTxs, countTxs, singleBlock } from './base-api';
 
 export const getListTxApi = async params => {
   let response = await _get(params, listTxs);
