+++ src/service/tweb3/transactions.js
@@ -1,6 +1,6 @@
 import tweb3 from '../../tweb3';
-import store from '../../store';
-import * as actions from '../../store/actions';
+// import store from '../../store';
+// import * as actions from '../../store/actions';
 
 export const getTxHistoryByAddress = async address => {
   try {
